backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "grandalf>=0.8",
    "httpx>=0.28.1",
    "langchain>=0.3.0",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-tavily>=0.2.17",
    "langgraph>=0.2.0",
    "networkx>=3.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "slowapi>=0.1.0",
    "sse-starlette>=2.0.0",
    "tavily-python>=0.5.0",
    "uvicorn>=0.30.0",
]
